+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Notes for Multi Translator
 
+## 2.9.0 - 2024-11-29
+
+### Added
+
+- on the fly settings override with "more options" button in the sidebar
+- translate to another target site from the sidebar
+
 ## 2.8.0 - 2024-11-15
 
 ### Added
@@ -111,6 +118,13 @@
 
 ### Updated
 - Craft 5 support
+
+## 1.14.0 - 2024-11-27
+
+### Added
+
+- on the fly settings override with "more options" button in the sidebar
+- translate to another target site from the sidebar
 
 ## 1.13.0 - 2024-11-15